requirements.txt
anthropic==0.3.11
google-cloud-aiplatform==1.32.0
openai==0.28.0
tenacity==8.2.3
python-dotenv==1.0.0
seaborn==0.12.2
pandas==2.1.0
ruamel.yaml==0.17.32